Abstract routes and controllers from server.js to their own files and folders.
