+++ server/routes/jobRouter.js
@@ -1,0 +1,15 @@
+import { Router } from "express";
+import {
+    getAllJobs,
+    getJob,
+    createJob,
+    updateJob,
+    deleteJob,
+} from "../controllers/jobController.js";
+
+const router = Router();
+
+router.route("/").get(getAllJobs).post(createJob);
+router.route("/:id").get(getJob).patch(updateJob).delete(deleteJob);
+
+export default router;
